Use the standard lexer
The contextual error is throwing errors intermittently. Switching
to the standard lexer fixes this but started raising other grammar
errors. Cleaning up the grammar and changing some rules to be terminals
fixes the standard lexer errors and allows us to use them.

Also fixed some places where the grammar was not following the hcl2
spec.